+++ src/bedrock/response.py
@@ -5,16 +5,8 @@ from typing import Any
 
 from attrs import define
 
+from .exceptions import CommandRequestError
 
-def parse(data: Mapping[str, Any]) -> CommandResponse | CommandResponseError:
-    """Parses a JSON object sent by the client."""
-    message = data["body"].get("statusMessage")
-    status=data["body"]["statusCode"]
-
-    res = CommandResponse(message, status)
-    if not res.ok():
-        return CommandResponseError(message, status)
-    return res
 
 @define
 class CommandResponse:
@@ -27,11 +19,16 @@ class CommandResponse:
     def message(self) -> str:
         """The message of the response."""
         return self._message
-    
+
     @property
     def status(self) -> int:
         """The status of the response."""
         return self._status
+
+    @property
+    def ok(self) -> bool:
+        """Returns ``True`` when the command has been executed successfully."""
+        return self.status == 0
 
     @classmethod
     def parse(cls, data: Mapping[str, Any]) -> CommandResponse:
@@ -41,31 +38,10 @@ class CommandResponse:
             status=data["body"]["statusCode"],
         )
 
-    def ok(self) -> bool:
-        """Returns ``True`` when the command has been executed successfully."""
-        return self.status == 0
-
-@define
-class CommandResponseError(Exception):
-    """A response sent by the client that indicates an error."""
-
-    _message: str
-    _status: int
-
-    @property
-    def message(self) -> str:
-        """The message of the response."""
-        return self._message
-    
-    @property
-    def status(self) -> int:
-        """The status of the response."""
-        return self._status
-
-    @classmethod
-    def parse(cls, data: Mapping[str, Any]) -> CommandResponseError:
-        """Parses a JSON object sent by the client."""
-        return cls(
-            message=data["body"].get("statusMessage"),
-            status=data["body"]["statusCode"],
-        )
+    def raise_for_status(self) -> None:
+        """
+        Raises :class:`bedrock.exceptions.CommandRequestError` if the command did
+        not run successfully. Otherwise this method returns ``None``.
+        """
+        if not self.ok:
+            raise CommandRequestError(self.message, self.status)